+++ src/defines/settings.h
@@ -29,35 +29,24 @@
 #define OIDC_PASSWORD_ENV_NAME "OIDC_ENCRYPTION_PW"
 
 // file names
-/**
- * the path to the config dir, if not provided via make
- */
 #define ISSUER_CONFIG_FILENAME "issuer.config"
 #define PUBCLIENTS_FILENAME "pubclients.config"
-#ifdef __MSYS__
 
-#define CONFIG_PATH_WIN "config/"
-#define ETC_ISSUER_CONFIG_FILE CONFIG_PATH_WIN "/" ISSUER_CONFIG_FILENAME
-#define ETC_PUBCLIENTS_CONFIG_FILE CONFIG_PATH_WIN "/" PUBCLIENTS_FILENAME
+#if defined __MINGW32__ || defined __MSYS__
+const char*                        CERT_FILE();
+const char*                        ETC_ISSUER_CONFIG_FILE();
+const char*                        ETC_PUBCLIENTS_CONFIG_FILE();
 
-#define CERT_PATH "/" CONFIG_PATH_WIN "ca-bundle.crt"
 #define OIDC_AGENT_REGISTRY "SOFTWARE\\oidc-agent"
-
+#define SOCKET_LOOPBACK_ADDRESS "127.0.0.1"
 #else
-
 #ifndef CONFIG_PATH
 #define CONFIG_PATH "/etc"
 #endif
+
 #define ETC_ISSUER_CONFIG_FILE CONFIG_PATH "/oidc-agent/" ISSUER_CONFIG_FILENAME
-#define PUBCLIENTS_FILENAME "pubclients.config"
 #define ETC_PUBCLIENTS_CONFIG_FILE \
   CONFIG_PATH "/oidc-agent/" PUBCLIENTS_FILENAME
-
-#endif
-
-#ifdef __MINGW32__
-#define OIDC_AGENT_REGISTRY "SOFTWARE\\oidc-agent"
-#define SOCKET_LOOPBACK_ADDRESS "127.0.0.1"
 #endif
 
 #define MAX_PASS_TRIES 3
@@ -76,7 +65,11 @@
 #define CONF_ENDPOINT_SUFFIX ".well-known/openid-configuration"
 #define OAUTH_CONF_ENDPOINT_SUFFIX ".well-known/oauth-authorization-server"
 
+#ifndef __MSYS__
+#ifndef __MINGW32__
 extern char* possibleCertFiles[4];
+#endif
+#endif
 
 /**
  * prefix for tmp-files generated during account generation;
@@ -84,19 +77,20 @@ extern char* possibleCertFiles[4];
  * saved in a file prefixed with that string
  */
 
+#define CLIENT_TMP_PREFIX \
+  "/tmp/oidc-gen:"  // This is not a filesystem path, this is only an oidc-agent
+                    // internal 'file'
 #ifdef __MSYS__
-#define CLIENT_TMP_PREFIX "tmp/oidc-gen:"
-#define AGENTDIR_LOCATION_CONFIG "config/"
-#define AGENTDIR_LOCATION_DOT "config/"
+#define AGENTDIR_LOCATION_CONFIG "$LOCALAPPDATA/oidc-agent/oidc-agent/"
+#define AGENTDIR_LOCATION_DOT "$userprofile/Documents/oidc-agent/"
 #else
-#define CLIENT_TMP_PREFIX "/tmp/oidc-gen:"
 #define AGENTDIR_LOCATION_CONFIG "~/.config/oidc-agent/"
 #define AGENTDIR_LOCATION_DOT "~/.oidc-agent/"
 #endif
 
 #ifdef __linux__
 #define URL_OPENER "xdg-open"
-#elif __APPLE__
+#else
 #define URL_OPENER "open"
 #endif